REFACTOR: 파일 경로 재구성
