+++ 토이_임준형/week7/api/src/main/java/week7/api/domain/domain/Review.java
@@ -1,4 +1,4 @@
-package week7.api.domain;
+package week7.api.domain.domain;
 
 import java.util.ArrayList;
 import java.util.List;
