+++ 토이_임준형/week7/api/src/main/java/week7/api/domain/domain/Status.java
@@ -1,4 +1,4 @@
-package week7.api.domain;
+package week7.api.domain.domain;
 
 public enum Status {
     NOT_STARTED, IN_PROGRESS, COMPLETED
